Create performance tests; add refactoring
