+++ cloud-formation/python/constants.py
@@ -1,0 +1,6 @@
+# Constants
+APIS = 'apis'
+SERVICES = 'services'
+IN_API = 'in-api'
+OUT_API = 'out-api'
+PROPERTIES = 'properties'
